Added class version of program
